refactor: move HomeScreen and googleSignin
Move HomeScreen to its own screen file within routes
Move googleSignin to be nested within hooks and use index.ts index.mock.ts
Update google user type to include missing parameters
Move google configuration to useEffect hook
