+++ App.tsx
@@ -1,68 +1,9 @@
-// In App.js in a new project
-
 import * as React from 'react';
 import {NavigationContainer} from '@react-navigation/native';
 import {createNativeStackNavigator} from '@react-navigation/native-stack';
-import {
-  View,
-  Text,
-  InputWithLabel,
-  Button,
-  GoogleSigninButton,
-  Touchable,
-} from 'src/components/common';
-import {signin as strings} from 'src/locales';
-import {useGoogleSignIn} from './src/hooks/googleSignin';
-import routes from 'src/routes';
-import testIds from './src/test-ids';
-
-export function HomeScreen() {
-  const [username, setUsername] = React.useState('');
-  const [password, setPassword] = React.useState('');
-  const {userInfo, googleSignIn, googleSignOut, isSigninInProgress} =
-    useGoogleSignIn();
-
-  return (
-    <View style={{flex: 1, alignItems: 'center', justifyContent: 'center'}}>
-      <InputWithLabel
-        label={strings.username}
-        labelTestID={testIds.page.signin.usernameLabel}
-        testID={testIds.page.signin.usernameInput}
-        value={username}
-        onChangeText={setUsername}
-      />
-      <InputWithLabel
-        label={strings.password}
-        labelTestID={testIds.page.signin.passwordLabel}
-        testID={testIds.page.signin.passwordInput}
-        value={password}
-        onChangeText={setPassword}
-      />
-      <Button
-        testID={testIds.page.signin.signinBtn}
-        onPress={() => {}}
-        title={strings.submit}
-        accessibilityLabel={strings.submit}
-      />
-      {userInfo.userInfo ? (
-        <Touchable
-          testID={testIds.page.signin.googleSignOutButton}
-          onPress={googleSignOut}>
-          <Text>Google sign out</Text>
-        </Touchable>
-      ) : (
-        <GoogleSigninButton
-          testID={testIds.page.signin.googleSignInButton}
-          style={{width: 192, height: 48}}
-          size={GoogleSigninButton.Size.Wide}
-          color={GoogleSigninButton.Color.Dark}
-          onPress={googleSignIn}
-          disabled={isSigninInProgress}
-        />
-      )}
-    </View>
-  );
-}
+import {View, Text} from 'src/components/common';
+import {HomeScreen} from 'src/routes';
+import {routes} from 'src/types';
 
 export function DetailsScreen() {
   return (
