+++ src/hooks/googleSignin/index.mock.ts
@@ -3,19 +3,6 @@ import {GoogleSignin} from '@react-native-google-signin/google-signin';
 import type {GoogleUserInfoType} from 'src/types';
 import mocks from 'e2e/mocks';
 
-GoogleSignin.configure({
-  scopes: [],
-  webClientId: '1234567890',
-  offlineAccess: true,
-  hostedDomain: '',
-  forceCodeForRefreshToken: true,
-  accountName: '',
-  iosClientId: '',
-  googleServicePlistPath: '',
-  openIdRealm: '',
-  profileImageSize: 120,
-});
-
 export const useGoogleSignIn = () => {
   const [userInfo, setUserInfo] = React.useState<GoogleUserInfoType>({
     userInfo: null,
@@ -24,7 +11,6 @@ export const useGoogleSignIn = () => {
     React.useState<boolean>(false);
 
   const googleSignIn = React.useCallback(async () => {
-    console.log('**LOG** tapped mock:');
     setIsSigninInProgress(true);
     setUserInfo({userInfo: mocks.googleUserMock});
     setIsSigninInProgress(false);
@@ -32,6 +18,21 @@ export const useGoogleSignIn = () => {
 
   const googleSignOut = React.useCallback(async () => {
     setUserInfo({userInfo: null});
+  }, []);
+
+  React.useEffect(() => {
+    GoogleSignin.configure({
+      scopes: [],
+      webClientId: '1234567890',
+      offlineAccess: true,
+      hostedDomain: '',
+      forceCodeForRefreshToken: true,
+      accountName: '',
+      iosClientId: '',
+      googleServicePlistPath: '',
+      openIdRealm: '',
+      profileImageSize: 120,
+    });
   }, []);
 
   return {googleSignIn, googleSignOut, userInfo, isSigninInProgress};
