+++ src/types/index.js
@@ -1,0 +1,2 @@
+export * from './routes';
+export * from './user';
