+++ src/types/routes.ts
@@ -1,0 +1,6 @@
+export enum routes {
+  Home = 'Home',
+  Details = 'Details',
+}
+
+export type HomeScreenProps = {};
